ChatBot/scripts/requirements.txt
requests
langchain-text-splitters
sentence-transformers
supabase
python-dotenv
